requirements.txt
streamlit
streamlit-extras
google-cloud-aiplatform
google-cloud-storage 
opencv-python 
google-cloud-vision
tqdm